refactor: 도메인 기반 프로젝트 구조로 변경
