+++ src/main/kotlin/com/relogging/server/domain/image/dto/ImageResponse.kt
@@ -1,0 +1,8 @@
+package com.relogging.server.domain.image.dto
+
+data class ImageResponse(
+    val id: Long,
+    val url: String,
+    val caption: String?,
+    val orderIndex: Int?,
+)
