+++ src/main/kotlin/com/relogging/server/domain/image/repository/ImageRepository.kt
@@ -1,0 +1,6 @@
+package com.relogging.server.domain.image.repository
+
+import com.relogging.server.domain.image.entity.Image
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ImageRepository : JpaRepository<Image, Long>
