+++ src/main/kotlin/com/relogging/server/domain/image/service/ImageService.kt
@@ -1,0 +1,10 @@
+package com.relogging.server.domain.image.service
+
+import org.springframework.web.multipart.MultipartFile
+
+interface ImageService {
+    fun saveImageFile(
+        file: MultipartFile,
+        savePath: String,
+    ): String
+}
