--- src/main/kotlin/com/relogging/server/dto/request/NewsArticleRequest.kt
@@ -1,16 +1,0 @@
-package com.relogging.server.dto.request
-
-import jakarta.validation.constraints.NotBlank
-import java.time.LocalDate
-
-data class NewsArticleRequest(
-    @field:NotBlank
-    val title: String?,
-    @field:NotBlank
-    val content: String?,
-    val aiSummary: String?,
-    val source: String?,
-    val author: String?,
-    val publishedAt: LocalDate?,
-    val imageCaption: String? = "",
-)
