--- src/main/kotlin/com/relogging/server/dto/response/ImageResponse.kt
@@ -1,8 +1,0 @@
-package com.relogging.server.dto.response
-
-data class ImageResponse(
-    val id: Long,
-    val url: String,
-    val caption: String?,
-    val orderIndex: Int?,
-)
