+++ src/main/kotlin/com/relogging/server/global/BaseEntity.kt
@@ -1,0 +1,21 @@
+package com.relogging.server.global
+
+import jakarta.persistence.Column
+import jakarta.persistence.EntityListeners
+import jakarta.persistence.MappedSuperclass
+import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedDate
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+
+@MappedSuperclass
+@EntityListeners(value = [AuditingEntityListener::class])
+abstract class BaseEntity {
+    @CreatedDate
+    @Column(nullable = false, updatable = false)
+    lateinit var createAt: LocalDateTime
+
+    @LastModifiedDate
+    @Column(nullable = false)
+    lateinit var updateAt: LocalDateTime
+}
