+++ src/main/kotlin/com/relogging/server/infrastructure/crawling/service/CrawlingService.kt
@@ -1,0 +1,5 @@
+package com.relogging.server.infrastructure.crawling.service
+
+interface CrawlingService {
+    fun crawlAndSave(): Int
+}
