+++ src/main/kotlin/com/relogging/server/infrastructure/crawling/service/PloggingEventCrawlingServiceImpl.kt
@@ -1,0 +1,10 @@
+package com.relogging.server.infrastructure.crawling.service
+
+import org.springframework.stereotype.Service
+
+@Service("ploggingEventCrawlingService")
+class PloggingEventCrawlingServiceImpl : CrawlingService {
+    override fun crawlAndSave(): Int {
+        TODO("Not yet implemented")
+    }
+}
