+++ src/main/kotlin/com/relogging/server/infrastructure/openai/service/OpenAiService.kt
@@ -1,0 +1,5 @@
+package com.relogging.server.infrastructure.openai.service
+
+interface OpenAiService {
+    fun aiSummary(content: String): String
+}
