--- src/main/kotlin/com/relogging/server/repository/ImageRepository.kt
@@ -1,6 +1,0 @@
-package com.relogging.server.repository
-
-import com.relogging.server.entity.Image
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface ImageRepository : JpaRepository<Image, Long>
