--- src/main/kotlin/com/relogging/server/service/crawling/CrawlingService.kt
@@ -1,5 +1,0 @@
-package com.relogging.server.service.crawling
-
-interface CrawlingService {
-    fun crawlAndSave(): Int
-}
