--- src/main/kotlin/com/relogging/server/service/crawling/PloggingEventCrawlingServiceImpl.kt
@@ -1,10 +1,0 @@
-package com.relogging.server.service.crawling
-
-import org.springframework.stereotype.Service
-
-@Service("ploggingEventCrawlingService")
-class PloggingEventCrawlingServiceImpl : CrawlingService {
-    override fun crawlAndSave(): Int {
-        TODO("Not yet implemented")
-    }
-}
