--- src/main/kotlin/com/relogging/server/service/image/ImageService.kt
@@ -1,10 +1,0 @@
-package com.relogging.server.service.image
-
-import org.springframework.web.multipart.MultipartFile
-
-interface ImageService {
-    fun saveImageFile(
-        file: MultipartFile,
-        savePath: String,
-    ): String
-}
