--- src/main/kotlin/com/relogging/server/service/openai/OpenAiService.kt
@@ -1,5 +1,0 @@
-package com.relogging.server.service.openai
-
-interface OpenAiService {
-    fun aiSummary(content: String): String
-}
